Only display oppfolgingsplan-link for arbeidsgiver
