+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/dialogmote-poc.iml" filepath="$PROJECT_DIR$/.idea/dialogmote-poc.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/dialogmote-frontend.iml" filepath="$PROJECT_DIR$/.idea/dialogmote-frontend.iml" />
     </modules>
   </component>
 </project>
